backend/pyproject.toml
[tool.poetry]
name = "friendflix-server"
version = "0.1.0"
description = "Backend Flask server for friendflix"
authors = ["Ofer Sadan <ofersadan85@gmail.com>"]
readme = "../README.md"
license = "../LICENSE"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
Flask-Cors = "^4.0.1"
Flask-JWT-Extended = "^4.6.0"
gunicorn = "^22.0.0"
httpx = "^0.27.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
pytest-watch = "^4.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-watch = "^4.2.0"
mypy = "^1.10.1"
types-psycopg2 = "^2.9.21.20240417"
types-flask-cors = "^4.0.0.20240523"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
cache_dir = "~/.cache/pytest"
addopts = "-vv" # --cov --cov-report=html"

[tool.mypy]
pretty = true
incremental = true
python_version = "3.12"
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
